ipc: move compact message processing to handler logic.
IPC compact messages are ABI specific. Move to ABI handler.
No code changes.
