+++ src/ipc/handler-ipc3.c
@@ -62,6 +62,12 @@
 #include <stddef.h>
 #include <stdint.h>
 
+#if CONFIG_CAVS && CAVS_VERSION >= CAVS_VERSION_1_8
+#include <ipc/header-intel-cavs.h>
+#include <cavs/drivers/sideband-ipc.h>
+#define CAVS_IPC_TYPE_S(x)		((x) & CAVS_IPC_TYPE_MASK)
+#endif
+
 #define iGS(x) ((x) & SOF_GLB_TYPE_MASK)
 #define iCS(x) ((x) & SOF_CMD_TYPE_MASK)
 
@@ -1373,6 +1379,46 @@ static int ipc_glb_test_message(uint32_t header)
 		tr_err(&ipc_tr, "ipc: unknown test header 0x%x", header);
 		return -EINVAL;
 	}
+}
+#endif
+
+#if CONFIG_CAVS && CAVS_VERSION >= CAVS_VERSION_1_8
+static struct sof_ipc_cmd_hdr *ipc_cavs_read_set_d0ix(uint32_t dr, uint32_t dd)
+{
+	struct sof_ipc_pm_gate *cmd = ipc_get()->comp_data;
+
+	cmd->hdr.cmd = SOF_IPC_GLB_PM_MSG | SOF_IPC_PM_GATE;
+	cmd->hdr.size = sizeof(*cmd);
+	cmd->flags = dd & CAVS_IPC_MOD_SETD0IX_BIT_MASK;
+
+	return &cmd->hdr;
+}
+
+/*
+ * Read a compact IPC message or return NULL for normal message.
+ */
+struct sof_ipc_cmd_hdr *ipc_compact_read_msg(void)
+{
+	struct sof_ipc_cmd_hdr *hdr;
+	uint32_t dr;
+	uint32_t dd;
+
+	dr = ipc_read(IPC_DIPCTDR);
+	dd = ipc_read(IPC_DIPCTDD);
+
+	/* if there is no cAVS module IPC in regs go the previous path */
+	if (!(dr & CAVS_IPC_MSG_TGT))
+		return mailbox_validate();
+
+	switch (CAVS_IPC_TYPE_S(dr)) {
+	case CAVS_IPC_MOD_SET_D0IX:
+		hdr = ipc_cavs_read_set_d0ix(dr, dd);
+		break;
+	default:
+		return NULL;
+	}
+
+	return hdr;
 }
 #endif
